completely restructure the app and finalise how to integrate other providers
