+++ lib/providers/gitlab-class.js
@@ -1,0 +1,16 @@
+/**
+ * Gitlab base class
+ */
+const EventEmitter = require('events');
+/**
+ * Main
+ */
+class GitlabHandler extends EventEmitter {
+  constructor(options) {
+    super();
+    this.options = options;
+  }
+  // @TODO
+}
+// Export
+module.exports = GitlabHandler;
