+++ lib/providers/index.js
@@ -1,7 +1,7 @@
 /**
  * Return the provider based on the config
  */
-const githubWebhookHandler = require('github-webhook-handler');
+const github = require('./github');
 const gitee = require('./gitee');
 const gitlab = require('./gitlab');
 /**
@@ -15,6 +15,6 @@ module.exports = provider => {
     case 'gitlab':
       return gitlab;
     default:
-      return githubWebhookHandler;
+      return github;
   }
 };
